pyproject.toml
[project]
name = "octoopt2"
version = "0.1.0"
description = "Home energy cost optimizer for GivEnergy + Octopus Agile + Mitsubishi Ecodan"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "crccheck>=1.3.1",
    "holidays>=0.93",
    "numpy>=2.4.4",
    "openmeteo-requests>=1.7.5",
    "pulp>=3.3.0",
    "pymelcloud>=2.11.0",
    "python-dotenv>=1.2.2",
    "requests-cache>=1.3.1",
    "retry-requests>=2.0.0",
    "solcast>=1.3.1",
]

[project.scripts]
octoopt2 = "octoopt2.main:main"
octoopt2-setup = "scripts.setup:main"
octoopt2-preload-consumption = "scripts.preload_consumption:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "hatchling>=1.29.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
